Created nonRenderableFiles, updated readme, fix testing functions
