+++ packages/templates/clients/websocket/test/integration-test/integration.test.js
@@ -1,8 +1,7 @@
-const { cleanTestResultPaths } = require('@asyncapi/generator-helpers');
 const path = require('path');
 const { readFile, stat } = require('fs').promises;
 const Generator = require('@asyncapi/generator');
-const { listFiles } = require('@asyncapi/generator-helpers');
+const { listFiles, cleanTestResultPaths } = require('@asyncapi/generator-helpers');
 const { runCommonTests } = require('./common-test.js');
 const asyncapi_v3_path_slack = path.resolve(__dirname, '../__fixtures__/asyncapi-slack-client.yml');
 const asyncapi_v3_path_hoppscotch = path.resolve(__dirname, '../__fixtures__/asyncapi-hoppscotch-client.yml');
